deploy/docker/memory-service/requirements.txt
# Titan Neural Memory Service Dependencies
# Reference: ADR-024 - Titan Neural Memory Architecture Integration

# Core ML/DL (PyTorch installed separately with CUDA)
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# gRPC for service communication
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Health checks and metrics
prometheus-client>=0.19.0
aiohttp>=3.9.0

# AWS integration
boto3>=1.34.0
botocore>=1.34.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
structlog>=24.1.0

# Async support
asyncio-throttle>=1.0.0

# Security
cryptography>=41.0.0
